Add image
We want to see whether accessing static resources also means that a
session is loaded. Intuition says yes, but data is better.

If a session is loaded to handle requests for static assets, then
we should probably look at whether we can reduce requests for static
assets.

This is also a good idea from a web performance optimisation view.

See https://developer.yahoo.com/performance/rules.html

A session *is* loaded by Spring Session JDBC for requests for the new
1x1.gif asset.

* Refactors the test scripts to extract another common function
* adds a new image in resources/static, which is a Spring Boot default
* configures debug logging on Spring Session